requirements.txt
ultralytics>=8.0.0
opencv-python-headless
numpy
tqdm
pillow
scipy
matplotlib